Fix null image and mandatory email
Fixing error which would not let you get into
your account as the blade requires image which
is not added in the account resultin in a null
object and preventing visits to your own profile.

Along with that removing the User class logic
which implements the MustVerifyEmail and
CanResetPassword interfaces in order to make
the smtp server non mandatory.

Signed-off-by: Martin Dekov <mvdekov@gmail.com>