Enable user segregation
Enabling user segregation from the registration
form which existed only as placeholder.

Migration was added to reflect whether a user
is organizer (true) or not (false means user is client)
and the form was augmented to reflect the change.

Now users can be organizers or not(clients). Tested
the change by adding two users.

Signed-off-by: Martin Dekov <mvdekov@gmail.com>