Add events view
Adding view for events which currently has
placeholdeholders and variables. It is not
possible to create events from it for now.

Signed-off-by: Martin Dekov <mvdekov@gmail.com>

Enable database organizers to be listed

Enabling the organizer users from the database
to be listed as targets agains which you
can request events.

Signed-off-by: Martin Dekov <mvdekov@gmail.com>

Add event creation

Adding ability to create events and restricting
event creation requests on for users which have
account either organizator or client

Signed-off-by: Martin Dekov <mvdekov@gmail.com>

Add one to many relationship between user and event

Adding semantic relationship between user and event.
The organizer user holds a list of events in json
format to approve.

The client user hold a list of events in json
format for events waiting to be approved.

Signed-off-by: Martin Dekov <mvdekov@gmail.com>

Add event approval interaction

Adding two views which take care of the
approved events and the waiting for
approval events. Currently the approving
itself happens manually through changing
the value in the DB down the line a button
which is now just a list will be added
to edit the event, aka approving it from
the organizer side

Signed-off-by: Martin Dekov <mvdekov@gmail.com>

Add approval mechanism

Adding mechanism which will make events approval
possible by the organizer. Now you don't have to
change the values manually in the database the
whole process of approval is done by an organizer
account

Signed-off-by: Martin Dekov <mvdekov@gmail.com>