Add final touches to images and comments
Adding final touches to images and comments milestone.
If image is not specified error will be returned instead
throwing an exception. Added flexbox for the images div in
order to go wrap on the next row when space ends. Added empty
storage folder with readme so you don't have an exception
telling you folder like that does not exist.

Signed-off-by: Martin Dekov <mvdekov@gmail.com>